+++ ATK/Utility/FFT.cpp
@@ -135,7 +135,7 @@ namespace ATK
 
     void process(const DataType_* input, std::size_t input_size, std::size_t size) const
     {
-      double factor = size;
+      double factor = static_cast<double>(size);
 #if ATK_USE_FFTW == 1
       for(std::size_t j = 0; j < std::min(input_size, size); ++j)
       {
@@ -221,7 +221,7 @@ namespace ATK
 
     void process(const std::complex<DataType_>* input, std::size_t input_size, std::size_t size) const
     {
-      double factor = size;
+      double factor = static_cast<double>(size);
 #if ATK_USE_FFTW == 1
       for (std::size_t j = 0; j < std::min(input_size, size); ++j)
       {
